+++ src/unoserver/converter.py
@@ -167,6 +167,22 @@ class UnoConverter:
             import_path, "_default", 0, input_props
         )
 
+        # Update document indexes
+        for ii in range(2):
+            # At first, update Table-of-Contents.
+            # ToC grows, so page numbers grow too.
+            # On second turn, update page numbers in ToC.
+            try:
+                document.refresh()
+                indexes = document.getDocumentIndexes()
+            except AttributeError:
+                # The document doesn't implement the XRefreshable and/or
+                # XDocumentIndexesSupplier interfaces
+                break
+            else:
+                for i in range(0, indexes.getCount()):
+                    indexes.getByIndex(i).update()
+
         # Now do the conversion
         try:
             # Figure out document type: